requirements.txt
pytest
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
python-dotenv==1.1.1
sqlparse==0.5.3
